Fix bug with clear and lighten/darken
Clear was setting all the empty squares to the value "white" instead of
clearing the value altogether. The lighten/darken functionality checks
for an empty value or expects an RGB value, so the value of "white" was
raising exceptions. Clear has been changed to set the value of all
squares to "" which lighten/darken appropriately ignores.